Extract runtime scoring module
