+++ backend/lined/.beads/interactions.jsonl
@@ -6,3 +6,4 @@
 {"id":"int-7ae524e7","kind":"field_change","created_at":"2026-06-01T14:19:13.525997Z","actor":"Oleksii Makieiev","issue_id":"lined-0zb","extra":{"field":"status","new_value":"closed","old_value":"in_progress","reason":"Completed calendar time-window refactor with validated window seam, conflict analyzer, header-bound requester checks, critic review, and ./gradlew check."}}
 {"id":"int-c971cfe3","kind":"field_change","created_at":"2026-06-03T21:58:03.353242Z","actor":"Oleksii Makieiev","issue_id":"lined-14m","extra":{"field":"status","new_value":"closed","old_value":"in_progress","reason":"Implemented runtime-aware scoring with local runtime-only output, baseline runtime scoring, SLO classification, tests, docs, and Notion sync planned for final handoff."}}
 {"id":"int-b4d0670f","kind":"field_change","created_at":"2026-06-03T22:07:31.48631Z","actor":"Oleksii Makieiev","issue_id":"lined-zdi","extra":{"field":"status","new_value":"closed","old_value":"in_progress","reason":"Addressed PR diff comments by extracting documented runtime fixtures/constants, adding node:test assertion plans, and refactoring runtime classification/scoring helpers; npm build/test and git diff --check pass."}}
+{"id":"int-50c9adc8","kind":"field_change","created_at":"2026-06-03T22:13:06.135446Z","actor":"Oleksii Makieiev","issue_id":"lined-jch","extra":{"field":"status","new_value":"closed","old_value":"in_progress","reason":"Moved runtime-aware SLO classification and score computation into runtimeScoring.ts so collectMetrics.ts remains collector orchestration; npm build/test pass."}}
